Create contract tests for mocks.
A missing piece of unit testing is the test between assumptions made in
one unit test and the functionality provided by the other unit. This is
the contract between two pieces of code.

td.when is great for making assumptions on another piece of code but
nothing makes sure that the assumptions a developer makes one day will
remain unbroken when the other piece of code changes. This commit
provides a td.contract that can be used to solidify the assumption made.

For example given the following function and it's unit test.
  const add = (a, b) => a + b
  assert(add(1, 2)).equal(3)  // All good :)

A developer might make the following assumption in another unit test.
  td.when(add(2, 3)).thenReturn(5)

If the implementation of add changes as follows.
  const lookupTable = { '12': 1 }
  const add = (a, b) => lookupTable[`${a}${b}`]
  assert(add(1, 2)).equal(3)  // All good :)

The unit test of add is happy, but the assumption is now wrong.
  td.when(add(2, 3)).thenReturn(5)  // False assumption :(

td.contract will in this case make sure the function does what
it's contract is promising to be.
  td.contract(add(2, 3)).thenReturn(5) // Logs a warning \o/

The point of td.contract is not to replace td.when. td.when is brilliant
for test driven development, but at some point, once the tests at hand
ar being finished, td.when can be converted to td.contract in order to
find problems when the contracts change in the future.

Caveats:
 - matchers do not work with td.contract